Added f_gem_node.cpp/.h/.inc for fortran using Fixed error overflowing in TValString::SetString() if fixed size. SD
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@805 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/submod/ms_multi_format.cpp
@@ -548,8 +548,8 @@ void TMulti::from_text_file_gemipm( const char *path )
 //   }
 //   else
 //  {
-    RoW = 0.99706137180;
-    EpsW = 78.245147705;
+    RoW = (float)0.99706137180;
+    EpsW = (float)78.245147705;
 //  }
 
 #ifndef IPMGEMPLUGIN
